ZKUI-152: Handle token expired xml error.
GetRolesForWebIdentity endpoint classical response is returned using json.
However, with the same accept: application/json header, in case of error the returned error is (or may be) formated using XML.
This is the case for the TokenExpired error response which is returned with 400 status code.
As the UI should display a re-login dialog when receiving this specific 400 response, we need to parse it.
Hence this try and fall response parsing mechanism that is super ugly :see_no_evil: