On Account side, this commit add assumable roles and preferredAssumable role features to the Account domain.
On Location side, I added support for locationStorageInfos to centralize the logic helping us to describe a
location in the UI (name + short type), it can be used together with useBucketConstraintLocation to display
the location of a bucket for example. Similarly on location side I added the cold and transient flags that
are necessary to help location name display and features.

On Bucket side, I added support to retrieve versionning information.